+++ device/libkernel_ga.c
@@ -56,7 +56,7 @@ uint64_t libkernel_ga (
 			float		DockConst_smooth,
 			uint		DockConst_num_of_intraE_contributors,
 			float		DockConst_grid_spacing,
-			uchar		DockConst_num_of_atypes,
+			uint		DockConst_num_of_atypes,
 			float		DockConst_coeff_elec,
 			float		DockConst_qasp,
 			float		DockConst_coeff_desolv,
